feat(js/api): rename useDatasetsSummary as useAssetsSummary
